market-spine-advanced/pyproject.toml
[project]
name = "market-spine-advanced"
version = "0.3.0"
description = "Advanced Market Spine with Celery, DLQ, external API ingestion, storage, and scheduling"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "structlog>=24.0",
    "ulid-py>=1.1",
    "click>=8.0",
    
    # Database
    "psycopg[binary,pool]>=3.1",
    
    # API
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
    
    # Task Queue
    "celery[redis]>=5.3",
    "redis>=5.0",
    
    # HTTP Client
    "httpx>=0.25",
    
    # Storage
    "boto3>=1.28",
    
    # Scheduling
    "croniter>=2.0",
    
    # Utilities
    "python-dotenv>=1.0",
    "tenacity>=8.2",
    "orjson>=3.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "httpx>=0.25",
    "moto[s3]>=4.0",
    "fakeredis>=2.20",
]

[project.scripts]
spine = "market_spine.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/market_spine"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
